+++ CHANGELOG.md
@@ -15,8 +15,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
  - Kit Editor: "Add Sample" now automatically resamples, normalizes and dithers the sample. No need to prepare samples using sox anymore.
  - Kit Editor: improved sound playback quality.
  - Kit Editor: when adding samples, blend wave frames to reduce impact of the [Game Boy wave refresh bug](https://www.devrs.com/gb/files/gbsnd3.gif).
- - Palette Editor: improved mid-tone generation.
  - Kit Editor: "Export sample" now always exports from ROM rather than intermediate 16-bit work sample.
+ - Kit Editor: Volume control now adjusts sample volume instead of prelisten volume.
+ - Palette Editor: improved mid-tone generation.
  - Various file dialog improvements.
  - Improved command line feedback.